requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.6.0

# Database
asyncpg>=0.29.0
pgvector==0.2.4

# AI/ML
openai==1.3.7
httpx<0.28

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
